arreglado login y hecho la bitacora
