+++ app/Models/Persona.php
@@ -8,6 +8,7 @@ use Illuminate\Database\Eloquent\Model;
 class Persona extends Model
 {
     use HasFactory;
+    
     protected $table = 'personas';
     protected $fillable = [
         'nombre',
